requirements.txt
aiohttp
RPi.GPIO
